refactor(domain): 도메인 객체들을 domain 패키지로 분리
