+++ src/main/java/lotto/LottoMain.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.*;
 import lotto.view.LottoInputView;
 import lotto.view.LottoOutputView;
