+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,11 +1,12 @@
-package lotto;
+package lotto.domain;
 
 import java.util.*;
 
+import lotto.domain.Ball;
+import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
