+++ src/test/java/lotto/domain/RankTest.java
@@ -1,7 +1,7 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.Rank;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
