+++ src/test/java/lotto/domain/WinningNumberSetTest.java
@@ -1,5 +1,8 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.Ball;
+import lotto.domain.Lotto;
+import lotto.domain.WinningNumberSet;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
